+++ spring-cloud-aws-test/src/test/java/io/awspring/cloud/test/sqs/BaseSqsIntegrationTest.java
@@ -1,0 +1,36 @@
+package io.awspring.cloud.test.sqs;
+
+import java.io.IOException;
+
+import org.junit.jupiter.api.BeforeAll;
+import org.testcontainers.containers.localstack.LocalStackContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+
+import static io.awspring.cloud.test.sqs.SqsSampleListener.QUEUE_NAME;
+import static org.testcontainers.containers.localstack.LocalStackContainer.Service.SQS;
+
+@Testcontainers
+abstract class BaseSqsIntegrationTest {
+
+	@Container
+	static LocalStackContainer localstack = new LocalStackContainer(
+			DockerImageName.parse("localstack/localstack:0.14.0")).withServices(SQS).withReuse(true);
+
+	@BeforeAll
+	static void beforeAll() throws IOException, InterruptedException {
+		// create needed queues in SQS
+		localstack.execInContainer("awslocal", "sqs", "create-queue", "--queue-name", QUEUE_NAME);
+	}
+
+	@DynamicPropertySource
+	static void registerSqsProperties(DynamicPropertyRegistry registry) {
+		// overwrite SQS endpoint with one provided by Localstack
+		registry.add("cloud.aws.sqs.endpoint", () -> localstack.getEndpointOverride(SQS).toString());
+	}
+
+}
